+++ sbai.ts
@@ -9,11 +9,7 @@ import {
 } from "@silverbulletmd/silverbullet/syscalls";
 import { decodeBase64 } from "https://deno.land/std@0.224.0/encoding/base64.ts";
 import { parse as parseYAML } from "https://deno.land/std@0.224.0/yaml/mod.ts";
-import {
-  getPageLength,
-  getSelectedTextOrNote,
-  updateFrontmatter,
-} from "./src/editorUtils.ts";
+import { getPageLength, updateFrontmatter } from "./src/editorUtils.ts";
 import type {
   EmbeddingModelConfig,
   ImageGenerationOptions,
@@ -42,6 +38,7 @@ import {
   folderName,
   queryObjects,
 } from "./src/utils.ts";
+import { installAICoreLibrary } from "./src/libraryInstaller.ts";
 
 /**
  * Similar to the above function, but meant for the config:loaded event.
@@ -53,6 +50,7 @@ export async function reloadConfig() {
     console.error("Failed to define config schemas:", error);
   }
   await initializeOpenAI(true);
+  await installAICoreLibrary();
 }
 
 /**
@@ -145,96 +143,6 @@ export async function selectEmbeddingModelFromConfig() {
     `Selected embedding model: ${selectedEmbeddingModelName}`,
   );
   console.log(`Selected embedding model:`, selectedEmbeddingModel);
-}
-
-/**
- * Prompts the user for a custom prompt to send to the LLM. If the user has text selected, the selected text is used as the note content.
- * If the user has no text selected, the entire note is used as the note content.
- * The response is streamed to the cursor position.
- */
-export async function callOpenAIwithNote() {
-  await initIfNeeded();
-  const selectedTextInfo = await getSelectedTextOrNote();
-  const userPrompt = await editor.prompt(
-    "Please enter a prompt to send to the LLM. Selected text or the entire note will also be sent as context.",
-  );
-  const noteName = await editor.getCurrentPage();
-  const currentDate = new Date();
-  const dateString = currentDate.toISOString().split("T")[0];
-  const dayString = currentDate.toLocaleDateString("en-US", {
-    weekday: "long",
-  });
-
-  await currentAIProvider.streamChatIntoEditor({
-    messages: [
-      {
-        role: "system",
-        content:
-          "You are an AI note assistant. Follow all user instructions and use the note context and note content to help follow those instructions. Use Markdown for any formatting.",
-      },
-      {
-        role: "user",
-        content:
-          `Note Context: Today is ${dayString}, ${dateString}. The current note name is "${noteName}".\nUser Prompt: ${userPrompt}\nNote Content:\n${selectedTextInfo.text}`,
-      },
-    ],
-    stream: true,
-  }, selectedTextInfo.to);
-}
-
-/**
- * Summarizes the selected text or the entire note if no text is selected.
- * Utilizes selected LLM to generate a summary.
- * Returns an object containing the summary and the selected text information.
- */
-export async function summarizeNote() {
-  await initIfNeeded();
-  const selectedTextInfo = await getSelectedTextOrNote();
-  console.log("selectedTextInfo", selectedTextInfo);
-  if (selectedTextInfo.text.length > 0) {
-    const noteName = await editor.getCurrentPage();
-    const response = await currentAIProvider.chatWithAI({
-      messages: [{
-        role: "user",
-        content:
-          `Please summarize this note using markdown for any formatting. Your summary will be appended to the end of this note, do not include any of the note contents yourself. Keep the summary brief. The note name is ${noteName}.\n\n${selectedTextInfo.text}`,
-      }],
-      stream: false,
-    });
-    console.log("OpenAI response:", response);
-    return {
-      summary: response,
-      selectedTextInfo: selectedTextInfo,
-    };
-  }
-  return { summary: "", selectedTextInfo: null };
-}
-
-/**
- * Uses a built-in prompt to ask the LLM for a summary of either the entire note, or the selected
- * text. Inserts the summary at the cursor's position.
- */
-export async function insertSummary() {
-  const { summary, selectedTextInfo } = await summarizeNote();
-  if (summary && selectedTextInfo) {
-    await editor.insertAtPos(
-      "\n\n" + summary,
-      selectedTextInfo.to,
-    );
-  }
-}
-
-/**
- * Uses a built-in prompt to ask the LLM for a summary of either the entire note, or the selected
- * text. Opens the resulting summary in a temporary right pane.
- */
-export async function openSummaryPanel() {
-  const { summary } = await summarizeNote();
-  if (summary) {
-    await editor.showPanel("rhs", 2, summary);
-  } else {
-    await editor.flashNotification("No summary available.");
-  }
 }
 
 /**
@@ -452,25 +360,6 @@ export async function enhanceNoteWithAI() {
   await tagNoteWithAI();
   await enhanceNoteFrontMatter();
   await suggestPageName();
-}
-
-/**
- * Streams a conversation with the LLM, inserting the responses at the cursor position as it is received.
- */
-export async function streamOpenAIWithSelectionAsPrompt() {
-  const selectedTextInfo = await getSelectedTextOrNote();
-  const cursorPos = selectedTextInfo.to;
-
-  await currentAIProvider.streamChatIntoEditor({
-    messages: [
-      {
-        role: "system",
-        content: "You are an AI note assistant in a markdown-based note tool.",
-      },
-      { role: "user", content: selectedTextInfo.text },
-    ],
-    stream: true,
-  }, cursorPos);
 }
 
 /**